backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
pyjwt
cryptography
requests
python-dotenv
langchain
langchain-community
langchain-google-genai
chromadb
tiktoken
pydantic
email-validator
jinja2
gspread
oauth2client
